+++ stripe_platform_interface/lib/src/models/payment_methods.dart
@@ -2,7 +2,6 @@ import 'package:freezed_annotation/freezed_annotation.dart';
 import 'package:json_annotation/json_annotation.dart';
 
 import 'address.dart';
-import 'card_field_input.dart';
 import 'payment_intents.dart';
 
 part 'payment_methods.freezed.dart';
@@ -263,7 +262,7 @@ enum PaymentMethodType {
 
 @Freezed(unionKey: 'type')
 class PaymentMethodParams with _$PaymentMethodParams {
- // const factory PaymentMethodParams(String type) = _PaymentMethodParamsDefault;
+  // const factory PaymentMethodParams(String type) = _PaymentMethodParamsDefault;
 
   @JsonSerializable(explicitToJson: true)
   @FreezedUnionValue('Card')