+++ stripe/CHANGELOG.md
@@ -1,3 +1,7 @@
-## 1.0.0+3
+## 1.0.1
+
+- Include Payment Sheet
+
+## 1.0.0
 
 * Initial beta release