+++ stripe_ios/CHANGELOG.md
@@ -1,3 +1,6 @@
-## 0.0.1
 
-* TODO: Describe initial release.
+## 0.0.1 - Payment Sheet
+
+Include Payment Sheet
+
+## 0.0.0 - First Version